refactor: split the HTTP and catalogue layers out of reconcile_models
reconcile_models.py had grown to 81 functions. The rate-limit-aware HTTP
client and the AST reader for OCS's own model catalogue are both leaf
layers - nothing in them reaches back into the pipeline - so they move to
reconcile_http.py and reconcile_catalogue.py, leaving 68.

A --dry-run against the live price table produces a byte-identical payload
before and after.
